Move MediaSmall to media module
