+++ app/src/main/kotlin/com/imashnake/animite/features/home/HomeScreen.kt
@@ -71,8 +71,8 @@ import com.imashnake.animite.core.extensions.bannerParallax
 import com.imashnake.animite.core.extensions.landscapeCutoutPadding
 import com.imashnake.animite.core.extensions.thenIf
 import com.imashnake.animite.core.ui.LocalPaddings
-import com.imashnake.animite.core.ui.MediaSmall
-import com.imashnake.animite.core.ui.MediaSmallRow
+import com.imashnake.animite.media.MediaSmall
+import com.imashnake.animite.media.MediaSmallRow
 import com.imashnake.animite.core.ui.ProgressIndicator
 import com.imashnake.animite.core.ui.layouts.BannerLayout
 import com.imashnake.animite.core.ui.layouts.TranslucentStatusBarLayout
@@ -84,7 +84,6 @@ import com.imashnake.animite.navigation.SharedContentKey.Component.Image
 import com.imashnake.animite.navigation.SharedContentKey.Component.Page
 import com.imashnake.animite.navigation.SharedContentKey.Component.Text
 import com.materialkolor.ktx.hasEnoughContrast
-import com.imashnake.animite.core.R as coreR
 import com.imashnake.animite.media.R as mediaR
 import com.imashnake.animite.navigation.R as navigationR
 
@@ -279,8 +278,6 @@ fun HomeRow(
                 image = media.coverImage,
                 label = media.title,
                 onClick = { onItemClicked(media) },
-                imageHeight = dimensionResource(coreR.dimen.media_image_height),
-                cardWidth = dimensionResource(coreR.dimen.media_card_width),
                 modifier = Modifier.sharedBounds(
                     rememberSharedContentState(
                         SharedContentKey(
