+++ app/src/main/kotlin/com/imashnake/animite/features/searchbar/SearchFrontDrop.kt
@@ -47,7 +47,7 @@ import com.imashnake.animite.api.anilist.type.MediaType
 import com.imashnake.animite.core.Constants
 import com.imashnake.animite.core.extensions.landscapeCutoutPadding
 import com.imashnake.animite.core.ui.LocalPaddings
-import com.imashnake.animite.core.ui.MediaSmall
+import com.imashnake.animite.media.MediaSmall
 import com.imashnake.animite.core.R as coreR
 import com.imashnake.animite.media.R as mediaR
 import com.imashnake.animite.navigation.R as navigationR
@@ -162,7 +162,7 @@ private fun SearchItem(
     Row(
         modifier = modifier
             .fillMaxWidth()
-            .clip(RoundedCornerShape(dimensionResource(coreR.dimen.media_card_corner_radius)))
+            .clip(RoundedCornerShape(dimensionResource(mediaR.dimen.media_card_corner_radius)))
             .clickable { onClick(item.id) }
     ) {
         MediaSmall(
