+++ media/build.gradle.kts
@@ -45,7 +45,7 @@ dependencies {
     implementation(libs.androidx.lifecycleRuntimeKtx)
 
     // Coil
-    implementation(libs.coil.compose)
+    implementation(libs.bundles.coil)
 
     // Compose
     implementation(libs.compose.animation)
