+++ media/src/main/res/values/dimens.xml
@@ -13,4 +13,10 @@
     <dimen name="character_image_height">137dp</dimen>
 
     <dimen name="trailer_corner_radius">30dp</dimen>
+
+    <dimen name="media_card_width">140dp</dimen>
+    <dimen name="media_image_height">200dp</dimen>
+    <dimen name="media_card_corner_radius">18dp</dimen>
+    <dimen name="media_card_text_padding_horizontal">16dp</dimen>
+    <dimen name="media_card_text_padding_vertical">10dp</dimen>
 </resources>
