+++ profile/src/main/kotlin/com/imashnake/animite/profile/tabs/Anime.kt
@@ -9,20 +9,18 @@ import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.res.dimensionResource
 import androidx.compose.ui.util.fastForEach
 import com.imashnake.animite.api.anilist.sanitize.profile.User
 import com.imashnake.animite.api.anilist.type.MediaType
 import com.imashnake.animite.core.ui.LocalPaddings
-import com.imashnake.animite.core.ui.MediaSmall
-import com.imashnake.animite.core.ui.MediaSmallRow
+import com.imashnake.animite.media.MediaSmall
+import com.imashnake.animite.media.MediaSmallRow
 import com.imashnake.animite.media.MediaPage
 import com.imashnake.animite.navigation.SharedContentKey
 import com.imashnake.animite.navigation.SharedContentKey.Component.Card
 import com.imashnake.animite.navigation.SharedContentKey.Component.Image
 import com.imashnake.animite.navigation.SharedContentKey.Component.Page
 import com.imashnake.animite.navigation.SharedContentKey.Component.Text
-import com.imashnake.animite.core.R as coreR
 
 @Composable
 fun AnimeTab(
@@ -80,8 +78,6 @@ private fun UserMediaList(
                                 )
                             )
                         },
-                        imageHeight = dimensionResource(coreR.dimen.media_image_height),
-                        cardWidth = dimensionResource(coreR.dimen.media_card_width),
                         modifier = Modifier.sharedBounds(
                             rememberSharedContentState(
                                 SharedContentKey(
